Draggable sdk panel cleanups
- remove `react-draggable` dep
- move `draggifyPanel` method to `lib/make-panel-draggable.js`
- remove `lib/get-document-dimensions.js`
- remove handling of `expand-panel` and `shrink-panel`
- remove draggable code from `components/app-view.js`
- fixes #167
- fixes #155
- fixes #154
- fixes #145
- fixes #149
- fixes #144
- fixes #140
