+++ index.js
@@ -4,16 +4,14 @@
  * http://mozilla.org/MPL/2.0/.
  */
 
+const system = require('sdk/system');
 const pageMod = require('sdk/page-mod');
 const getYouTubeUrl = require('./lib/get-youtube-url.js');
 const getVimeoUrl = require('./lib/get-vimeo-url.js');
 const launchVideo = require('./lib/launch-video');
 const sendMetricsData = require('./lib/send-metrics-data.js');
-const getDocumentDimensions = require('./lib/get-document-dimensions.js');
 const initContextMenuHandlers = require('./lib/context-menu-handlers.js');
-const { getActiveView } = require('sdk/view/core');
-
-let dimensions = getDocumentDimensions();
+const makePanelDraggable = require('./lib/make-panel-draggable.js');
 
 const panel = require('sdk/panel').Panel({
   contentURL: './default.html',
@@ -26,47 +24,10 @@ const panel = require('sdk/panel').Panel({
   }
 });
 
-// Makes an SDK panel draggable. Pass in an SDK panel.
-function draggifyPanel(panel) {
-  // Remove the panel from the XUL DOM, make some attribute changes, then
-  // reattach it. Reseating in the DOM triggers updates in the XBL bindings
-  // that give the panel draggability and remove some SDK styling.
-  const panelEl = getActiveView(panel);
-  const parentEl = panelEl.parentNode;
-
-  parentEl.removeChild(panelEl);
-
-  panelEl.setAttribute('noautohide', true);
-  panelEl.setAttribute('backdrag', true);
-  panelEl.setAttribute('style', '-moz-appearance: none; border: 0; margin: 0; background: rgba(0,0,0,0)');
-  panelEl.removeAttribute('type');
-
-  // Next, we need a XUL document to create a drag handle. There may be better
-  // ways to obtain the document element, but this works:
-  let doc = parentEl;
-  while (doc !== null && doc.nodeType !== 9) {
-    doc = doc.parentNode;
-  }
-
-  const dragHandle = doc.createElement('label');
-  dragHandle.id = 'backdragspot';
-  // TODO: the drag handle doesn't seem to be visible--maybe the iframe is set
-  // to full height, or has a higher z-index?
-  dragHandle.setAttribute('value', 'click here to drag the thing');
-  dragHandle.setAttribute('style', 'background: #2b2b2b; border: 1px solid black; color: #d5d5d5; cursor: grab');
-  dragHandle.setAttribute('hidden', true);
-  dragHandle.onmousedown = () => { dragHandle.style.cursor = 'grabbing' }
-  dragHandle.onmouseup = () => { dragHandle.style.cursor = 'grab' }
-  panelEl.appendChild(dragHandle);
-
-  // make the drag handle only visible on mouseover
-  panelEl.onmouseenter = () => { dragHandle.setAttribute('hidden', false) };
-  panelEl.onmouseleave = () => { dragHandle.setAttribute('hidden', true) };
-
-  parentEl.appendChild(panelEl);
+// check for linux here
+if (system.platform !== 'linux') {
+  makePanelDraggable(panel);
 }
-
-draggifyPanel(panel);
 
 panel.port.on('addon-message', opts => {
   const title = opts.action;
@@ -101,20 +62,6 @@ panel.port.on('addon-message', opts => {
         left: 10
       }
     });
-  } else if (title === 'expand-panel') {
-    panel.hide();
-    panel.show({
-      width: dimensions.width,
-      height: dimensions.height,
-      position: opts.position
-    });
-  } else if (title === 'shrink-panel') {
-    panel.hide();
-    panel.show({
-      width: opts.style.width,
-      height: opts.style.height,
-      position: opts.position
-    });
   } else if (title === 'metrics-event') {
     sendMetricsData(opts, panel);
   }
@@ -145,9 +92,6 @@ pageMod.PageMod({
         panel.hide();
         panel.show();
       }
-
-      // update our document dimensions
-      dimensions = getDocumentDimensions();
     });
   }
 });
