+++ package.json
@@ -50,7 +50,6 @@
     "pre-commit": "1.1.3",
     "react": "15.1.0",
     "react-dom": "15.1.0",
-    "react-draggable": "2.2.1",
     "react-tooltip": "3.0.13",
     "watchify": "3.7.0"
   }
